adds frontend/backend for comments
adds comment model and database migration
adjusts track with comment model
adds add comment, show comments, and delete comments to leaderboardController
adds route for the above
adds frontend section on dashboard for comments to be expanded
adds a view for comments
adjusts comment views with access level for deletions

+++ project/app/Actions/Jetstream/DeleteUser.php
@@ -2,9 +2,10 @@
 
 namespace App\Actions\Jetstream;
 
+use App\Models\Comment;
+use App\Models\Party;
 use App\Models\Rating;
 use App\Models\Track;
-use Illuminate\Support\Facades\Auth;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class DeleteUser implements DeletesUsers
@@ -17,15 +18,32 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
+        // get the users' ratings, delete them all, and adjust the track accordingly
         $ratings = Rating::query()->where('user_id', '=', $user->id)->get();
-
-        foreach($ratings as $rating) {
-            $track = Track::query()->where('id', '=', $rating->track->id)->first();
-            $track->rating = $track->rating - 1;
-            $track->save();
-            $rating->delete();
+        if ($ratings != null) {
+            foreach ($ratings as $rating) {
+                $track = Track::query()->where('id', '=', $rating->track->id)->first();
+                $track->rating = $track->rating - 1;
+                $track->save();
+                $rating->delete();
+            }
         }
 
+        // get the users' comments and delete them all
+        $comments = Comment::query()->where('user_id', '=', $user->id)->get();
+        if ($comments != null) {
+            foreach ($comments as $comment) {
+                $comment->delete();
+            }
+        }
+
+        // get the users' party and delete it
+        $party = Party::query()->where('user_id', '=', $user->id)->first();
+        if ($party != null) {
+            $party->delete();
+        }
+
+        // delete the users' photo, tokens, accounts
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->connectedAccounts->each->delete();
